refactor: Seclude `toProtobuTags` utility
