+++ node/packages/sdk/lib/trace-span.js
@@ -12,35 +12,10 @@ const { AsyncLocalStorage } = require('async_hooks');
 const ensureSpanName = require('./get-ensure-resource-name')('INVALID_TRACE_SPAN_NAME');
 const emitter = require('./emitter');
 const Tags = require('./tags');
-const toLong = require('./to-long');
+const toProtobufTags = require('./to-protobuf-tags');
 const generateId = require('./generate-id');
 const resolveEpochTimestampString = require('./resolve-epoch-timestamp-string');
 const toProtobufEpochTimestamp = require('./to-protobuf-epoch-timestamp');
-
-const resolvePorotbufValue = (key, value) => {
-  switch (key) {
-    // enum cases
-    case 'aws.lambda.outcome':
-      switch (value) {
-        case 'success':
-          return 1;
-        case 'error:handled':
-          return 5;
-        default:
-          // Will error in tests
-          return null;
-      }
-    default:
-      if (Array.isArray(value)) {
-        if (typeof value[0] === 'number') return value.map(toLong);
-        return value;
-      }
-      return typeof value === 'number' ? toLong(value) : value;
-  }
-};
-
-const snakeToCamelCase = (string) =>
-  string.replace(/_(.)/g, (ignore, letter) => letter.toUpperCase());
 
 class StringifiableSet extends Set {
   toJSON() {
@@ -185,17 +160,6 @@ class TraceSpan {
     };
   }
   toProtobufObject() {
-    const tags = {};
-    for (const [key, value] of this.tags) {
-      let context = tags;
-      const keyTokens = key.split('.').map((token) => snakeToCamelCase(token));
-      const lastToken = keyTokens.pop();
-      for (const token of keyTokens) {
-        if (!context[token]) context[token] = {};
-        context = context[token];
-      }
-      context[lastToken] = resolvePorotbufValue(key, value);
-    }
     return {
       id: Buffer.from(this.id),
       traceId: Buffer.from(this.traceId),
@@ -205,7 +169,7 @@ class TraceSpan {
       endTimeUnixNano: this.endTime ? toProtobufEpochTimestamp(this.endTime) : undefined,
       input: this.input || undefined,
       output: this.output || undefined,
-      tags,
+      tags: toProtobufTags(this.tags),
     };
   }
   get spans() {
